Update 06-dispatching-to-a-class.md
Am new to composer. When I try running the project, a CLASS NOT FOUND error was thrown.

I had to do a few search on StackOverflow to get it working. 

New users like me will forget to dump-update before running.  So I suggest you add a note to fix this small problem.